Update form component, validation, id assignment and cleaning
Add Change handlers for storing the information that is written inside the fields of the form.
Create an state for appointment to store the values of the fields of the form.
Validate all the fields inside the form with an error state.
Assign a unique id for each appointment with uuid.
Clean all fields, when the appointment is submitted.
Create a state in app component to keep track of all the appointments made.